Split readonly tests into separate test file and add error cases
